+++ handlers/harvest_api_v1/test/get_input.rb
@@ -1,0 +1,14 @@
+{
+  'info' => {
+    'account_id' => "",
+    'access_token' => "",
+    'api_location' => "https://api.harvestapp.com/v2",
+    'enable_debug_logging' => 'true'
+  },
+  'parameters' => {
+    'error_handling' => 'Error Message',
+    'method' => 'POST',
+    'path' => '/projects',
+    'body' => ''
+  }
+}